requirements.txt
streamlit>=1.32.0
python-docx>=1.1.0
python-pptx>=0.6.23
PyMuPDF>=1.24.0
openai==0.28.1
nltk>=3.8.1
rouge-score>=0.1.2
pandas>=2.2.0
